Convert serialization framework into something similar to the templating framework
