+++ schemacrawler-scripting/src/main/resources/META-INF/services/schemacrawler.tools.executable.CommandProvider
@@ -1,3 +1,3 @@
 schemacrawler.tools.integration.script.ScriptCommandProvider
 schemacrawler.tools.integration.template.TemplateCommandProvider
-schemacrawler.tools.integration.serialization.SerializationCommandProvider
+schemacrawler.tools.integration.serialize.SerializationCommandProvider
