+++ schemacrawler-scripting/src/test/java/schemacrawler/test/script/SchemaCrawlerExecutableChainTest.java
@@ -26,7 +26,7 @@ http://www.gnu.org/licenses/
 ========================================================================
 */
 
-package schemacrawler.test;
+package schemacrawler.test.script;
 
 
 import static org.hamcrest.MatcherAssert.assertThat;
