+++ schemacrawler-scripting/src/test/java/schemacrawler/test/script/ScriptingLanguageTest.java
@@ -26,60 +26,55 @@ http://www.gnu.org/licenses/
 ========================================================================
 */
 
-package schemacrawler.test;
+package schemacrawler.test.script;
 
 
 import static org.hamcrest.MatcherAssert.assertThat;
+import static schemacrawler.test.utility.ExecutableTestUtility.executableExecution;
+import static schemacrawler.test.utility.ExecutableTestUtility.executableOf;
 import static schemacrawler.test.utility.FileHasContent.*;
-import static schemacrawler.test.utility.ScriptTestUtility.commandLineTemplateExecution;
+import static schemacrawler.test.utility.TestUtility.copyResourceToTempFile;
+
+import java.nio.file.Path;
+import java.sql.Connection;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import schemacrawler.test.utility.DatabaseConnectionInfo;
+import schemacrawler.schemacrawler.Config;
 import schemacrawler.test.utility.TestAssertNoSystemErrOutput;
 import schemacrawler.test.utility.TestAssertNoSystemOutOutput;
 import schemacrawler.test.utility.TestDatabaseConnectionParameterResolver;
+import schemacrawler.tools.executable.SchemaCrawlerExecutable;
 
 @ExtendWith(TestAssertNoSystemErrOutput.class)
 @ExtendWith(TestAssertNoSystemOutOutput.class)
 @ExtendWith(TestDatabaseConnectionParameterResolver.class)
-public class TemplatingCommandlineTest
+public class ScriptingLanguageTest
 {
 
-  @Test
-  public void commandlineFreeMarker(final DatabaseConnectionInfo connectionInfo)
+  private static Path executableScriptFromFile(final Connection connection,
+                                               final String language,
+                                               final Path scriptFile)
     throws Exception
   {
-    assertThat(outputOf(commandLineTemplateExecution(connectionInfo,
-                                                     "/plaintextschema.ftl")),
-               hasSameContentAs(classpathResource("executableForFreeMarker.txt")));
+    final SchemaCrawlerExecutable executable = executableOf("script");
+    final Config additionalConfiguration = new Config();
+    additionalConfiguration.put("script", scriptFile.toString());
+    additionalConfiguration.put("script-language", language);
+    executable.setAdditionalConfiguration(additionalConfiguration);
+
+    return executableExecution(connection, executable, "text");
   }
 
   @Test
-  public void commandlineMustache(final DatabaseConnectionInfo connectionInfo)
+  public void executableGroovy(final Connection connection)
     throws Exception
   {
-    assertThat(outputOf(commandLineTemplateExecution(connectionInfo,
-                                                     "/plaintextschema.mustache")),
-               hasSameContentAs(classpathResource("executableForMustache.txt")));
-  }
-
-  @Test
-  public void commandlineThymeleaf(final DatabaseConnectionInfo connectionInfo)
-    throws Exception
-  {
-    assertThat(outputOf(commandLineTemplateExecution(connectionInfo,
-                                                     "/plaintextschema.thymeleaf")),
-               hasSameContentAs(classpathResource("executableForThymeleaf.txt")));
-  }
-
-  @Test
-  public void commandlineVelocity(final DatabaseConnectionInfo connectionInfo)
-    throws Exception
-  {
-    assertThat(outputOf(commandLineTemplateExecution(connectionInfo,
-                                                     "/plaintextschema.vm")),
-               hasSameContentAs(classpathResource("executableForVelocity.txt")));
+    final Path scriptFile = copyResourceToTempFile("/plaintextschema.groovy");
+    assertThat(outputOf(executableScriptFromFile(connection,
+                                                 "groovy",
+                                                 scriptFile)),
+               hasSameContentAs(classpathResource("script_output.txt")));
   }
 
 }
