+++ schemacrawler-scripting/src/test/java/schemacrawler/test/template/TemplateLanguageTest.java
@@ -26,7 +26,7 @@ http://www.gnu.org/licenses/
 ========================================================================
 */
 
-package schemacrawler.test;
+package schemacrawler.test.template;
 
 
 import static org.hamcrest.CoreMatchers.is;
