+++ schemacrawler-scripting/src/test/java/schemacrawler/test/template/TemplatingCommandlineTest.java
@@ -26,54 +26,59 @@ http://www.gnu.org/licenses/
 ========================================================================
 */
 
-package schemacrawler.test;
+package schemacrawler.test.template;
 
 
 import static org.hamcrest.MatcherAssert.assertThat;
-import static schemacrawler.test.utility.ExecutableTestUtility.executableExecution;
-import static schemacrawler.test.utility.ExecutableTestUtility.executableOf;
 import static schemacrawler.test.utility.FileHasContent.*;
-import static schemacrawler.test.utility.TestUtility.copyResourceToTempFile;
-
-import java.nio.file.Path;
-import java.sql.Connection;
+import static schemacrawler.test.utility.ScriptTestUtility.commandLineTemplateExecution;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import schemacrawler.schemacrawler.Config;
+import schemacrawler.test.utility.DatabaseConnectionInfo;
 import schemacrawler.test.utility.TestAssertNoSystemErrOutput;
 import schemacrawler.test.utility.TestAssertNoSystemOutOutput;
 import schemacrawler.test.utility.TestDatabaseConnectionParameterResolver;
-import schemacrawler.tools.executable.SchemaCrawlerExecutable;
 
 @ExtendWith(TestAssertNoSystemErrOutput.class)
 @ExtendWith(TestAssertNoSystemOutOutput.class)
 @ExtendWith(TestDatabaseConnectionParameterResolver.class)
-public class TemplatingLanguageTest
+public class TemplatingCommandlineTest
 {
 
-  private static Path executableScriptFromFile(final Connection connection,
-                                               final String language,
-                                               final Path scriptFile)
+  @Test
+  public void commandlineFreeMarker(final DatabaseConnectionInfo connectionInfo)
     throws Exception
   {
-    final SchemaCrawlerExecutable executable = executableOf("template");
-    final Config additionalConfiguration = new Config();
-    additionalConfiguration.put("template", scriptFile.toString());
-    additionalConfiguration.put("templating-language", language);
-    executable.setAdditionalConfiguration(additionalConfiguration);
-
-    return executableExecution(connection, executable, "text");
+    assertThat(outputOf(commandLineTemplateExecution(connectionInfo,
+                                                     "/plaintextschema.ftl")),
+               hasSameContentAs(classpathResource("executableForFreeMarker.txt")));
   }
 
   @Test
-  public void executableGroovy(final Connection connection)
+  public void commandlineMustache(final DatabaseConnectionInfo connectionInfo)
     throws Exception
   {
-    final Path scriptFile = copyResourceToTempFile("/plaintextschema.vm");
-    assertThat(outputOf(executableScriptFromFile(connection,
-                                                 "velocity",
-                                                 scriptFile)),
+    assertThat(outputOf(commandLineTemplateExecution(connectionInfo,
+                                                     "/plaintextschema.mustache")),
+               hasSameContentAs(classpathResource("executableForMustache.txt")));
+  }
+
+  @Test
+  public void commandlineThymeleaf(final DatabaseConnectionInfo connectionInfo)
+    throws Exception
+  {
+    assertThat(outputOf(commandLineTemplateExecution(connectionInfo,
+                                                     "/plaintextschema.thymeleaf")),
+               hasSameContentAs(classpathResource("executableForThymeleaf.txt")));
+  }
+
+  @Test
+  public void commandlineVelocity(final DatabaseConnectionInfo connectionInfo)
+    throws Exception
+  {
+    assertThat(outputOf(commandLineTemplateExecution(connectionInfo,
+                                                     "/plaintextschema.vm")),
                hasSameContentAs(classpathResource("executableForVelocity.txt")));
   }
