+++ schemacrawler-scripting/src/test/java/schemacrawler/test/template/TemplatingTest.java
@@ -26,55 +26,59 @@ http://www.gnu.org/licenses/
 ========================================================================
 */
 
-package schemacrawler.test;
+package schemacrawler.test.template;
 
 
 import static org.hamcrest.MatcherAssert.assertThat;
-import static schemacrawler.test.utility.ExecutableTestUtility.executableExecution;
-import static schemacrawler.test.utility.ExecutableTestUtility.executableOf;
 import static schemacrawler.test.utility.FileHasContent.*;
-import static schemacrawler.test.utility.TestUtility.copyResourceToTempFile;
+import static schemacrawler.test.utility.ScriptTestUtility.templateExecution;
 
-import java.nio.file.Path;
 import java.sql.Connection;
 
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
-import schemacrawler.schemacrawler.Config;
 import schemacrawler.test.utility.TestAssertNoSystemErrOutput;
 import schemacrawler.test.utility.TestAssertNoSystemOutOutput;
 import schemacrawler.test.utility.TestDatabaseConnectionParameterResolver;
-import schemacrawler.tools.executable.SchemaCrawlerExecutable;
 
 @ExtendWith(TestAssertNoSystemErrOutput.class)
 @ExtendWith(TestAssertNoSystemOutOutput.class)
 @ExtendWith(TestDatabaseConnectionParameterResolver.class)
-public class ScriptingLanguageTest
+public class TemplatingTest
 {
 
-  private static Path executableScriptFromFile(final Connection connection,
-                                               final String language,
-                                               final Path scriptFile)
+  @Test
+  public void executableFreeMarker(final Connection connection)
     throws Exception
   {
-    final SchemaCrawlerExecutable executable = executableOf("script");
-    final Config additionalConfiguration = new Config();
-    additionalConfiguration.put("script", scriptFile.toString());
-    additionalConfiguration.put("script-language", language);
-    executable.setAdditionalConfiguration(additionalConfiguration);
-
-    return executableExecution(connection, executable, "text");
+    assertThat(outputOf(templateExecution(connection, "/plaintextschema.ftl")),
+               hasSameContentAs(classpathResource("executableForFreeMarker.txt")));
   }
 
   @Test
-  public void executableGroovy(final Connection connection)
+  public void executableMustache(final Connection connection)
     throws Exception
   {
-    final Path scriptFile = copyResourceToTempFile("/plaintextschema.groovy");
-    assertThat(outputOf(executableScriptFromFile(connection,
-                                                 "groovy",
-                                                 scriptFile)),
-               hasSameContentAs(classpathResource("script_output.txt")));
+    assertThat(outputOf(templateExecution(connection,
+                                          "/plaintextschema.mustache")),
+               hasSameContentAs(classpathResource("executableForMustache.txt")));
+  }
+
+  @Test
+  public void executableThymeleaf(final Connection connection)
+    throws Exception
+  {
+    assertThat(outputOf(templateExecution(connection,
+                                          "/plaintextschema.thymeleaf")),
+               hasSameContentAs(classpathResource("executableForThymeleaf.txt")));
+  }
+
+  @Test
+  public void executableVelocity(final Connection connection)
+    throws Exception
+  {
+    assertThat(outputOf(templateExecution(connection, "/plaintextschema.vm")),
+               hasSameContentAs(classpathResource("executableForVelocity.txt")));
   }
 
 }
