+++ schemacrawler-scripting/src/test/resources/commandline_help_output/help.serialize.stdout.txt
@@ -1,0 +1,9 @@
+** Create an offline catalog snapshot
+
+      --serialization-format=<serialization-format>
+         Specifies serialization output format
+         Optional, defaults to Java serialization
+
+         Use one of unknown, java, json
+
+
